requirements.txt
asgiref==3.8.1
channels==4.2.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg==3.2.3
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.9.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
